refactor(web): split cockpit domains out of main
